+++ ElvUI_MerathilisUI/Core/Changelog/7.32.lua
@@ -2,7 +2,10 @@ local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[732] = {
 	RELEASE_DATE = "TBD",
-	FIXES = {},
+	FIXES = {
+		'[Actionbars]: Fixed SpecBar not showing a distinct border color when Loot Specialization is set to "Current Specialization"',
+		"[Actionbars]: Fixed SpecBar's border color reverting to the default color shortly after login/reload",
+	},
 	NEW = {},
 	IMPROVEMENTS = {},
 }